+++ docs/docs/user-guide/link/included.py
@@ -1,0 +1,5 @@
+# Script can be included.
+
+
+def func(x, y):
+    return x + y