+++ tests/plugins/test_plugins_mkdocs.py
@@ -66,8 +66,8 @@ def test_plugins_mkdocs_plugin(plugin):
 
 
 def test_plugins_mkdocs_files(files, plugin, nav):
-    assert len(files.documentation_pages()) == 3
-    assert len(plugin.converter.pages) == 2
+    assert len(files.documentation_pages()) == 6
+    assert len(plugin.converter.pages) == 3
 
 
 def test_plugins_mkdocs_page_render(files, plugin, config):