+++ tests/renderers/embed/test_embed_parse_inline_code.py
@@ -15,22 +15,23 @@ def parse(parser):
 
 @pytest.fixture()
 def file(tmpdir, parse):
-    def func(path, source):
+    def func(path, content, source):
         path_ = tmpdir.join(path)
-        path_.write(source)
+        path_.write(content)
         current = os.path.abspath(os.path.curdir)
         os.chdir(tmpdir)
-        output = parse(path)
+        output = parse(source)
         os.chdir(current)
         return output
 
     return func
 
 
-def test_code_parse_file(file):
+def test_embed_parse_file(file):
     path = "hello.py"
-    source = "def func(x):\n    return 2 * x\n\nprint(f(3))\n"
-    output = file(path, source)
+    content = "def func(x):\n    return 2 * x\n\nprint(f(3))\n"
+    source = "=hello.py"
+    output = file(path, content, source)
     answer = (
         '\n<div class="source"><pre>'
         '<code class="python">def func(x):\n    return 2 * x\n\nprint(f(3))'
@@ -39,13 +40,13 @@ def test_code_parse_file(file):
     assert output == answer
 
 
-def test_code_parse_file_not_founed(parse):
+def test_embed_parse_file_not_founed(parse):
     path = "xxx.py"
     output = parse(path)
-    assert output == '<p style="font-color:red">File not found: xxx.py</p>\n'
+    assert '<p style="font-color:red">File not found:' in output
 
 
-def test_code_parse_inspect(parse):
+def test_embed_parse_inspect(parse):
     execute("import pheasant")
     output = parse("?pheasant")
     answer = '\n<div class="source">' '<pre><code class="python">__version__ ='