+++ tests/renderers/jupyter/test_jupyter_client.py
@@ -60,3 +60,14 @@ def test_kernel_manager():
     assert not km.is_alive()
     km = get_kernel_manager("python")
     assert km.is_alive()
+
+
+def test_stream_joinner():
+    outputs = execute("import sys\nsys.stdout.write('1')\nsys.stdout.write('\x082')\n1")
+    assert outputs[0]['text'] == '2'
+    assert outputs[1]['data']['text/plain'] == '1'
+
+    source = "sys.stdout.write('1')\nsys.stdout.write('2')\nsys.stderr.write('3')\n1"
+    outputs = execute(source)
+    assert outputs[0]["text"] == '12'
+    assert outputs[1]["text"] == '3'