+++ docs/docs/about/release-notes.md
@@ -14,10 +14,15 @@ You can determine your currently installed version using `pheasant --version`:
 
 ~~~
 $ pheasant --version
-pheasant, version 2.0.13 from /path/to/pheasant (Python 3.7)
+pheasant, version 2.1.0 from /path/to/pheasant (Python 3.7)
 ~~~
 
 ## Change log
+
+### Version 2.1.0 (2019-04-04)
+
+* Add feature to import extra pages.
+* Jupyter: multi-stdout/err into a joined cell.
 
 ### Version 2.0.13 (2019-04-04)